rename lib name since dash is not pythonic
